Exctract sites C[!R]UD operations into SitesEndpoint
