+++ src/main/java/org/ccjmne/orca/api/rest/admin/UpdateEndpoint.java
@@ -1,32 +1,24 @@
 package org.ccjmne.orca.api.rest.admin;
 
 import static org.ccjmne.orca.jooq.classes.Tables.EMPLOYEES;
-import static org.ccjmne.orca.jooq.classes.Tables.SITES;
 import static org.ccjmne.orca.jooq.classes.Tables.SITES_EMPLOYEES;
-import static org.ccjmne.orca.jooq.classes.Tables.SITES_TAGS;
 import static org.ccjmne.orca.jooq.classes.Tables.UPDATES;
 import static org.ccjmne.orca.jooq.classes.Tables.USERS;
 
 import java.text.ParseException;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
 import javax.inject.Inject;
 import javax.ws.rs.Consumes;
-import javax.ws.rs.DELETE;
 import javax.ws.rs.ForbiddenException;
 import javax.ws.rs.POST;
-import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
 import javax.ws.rs.core.MediaType;
 
-import org.apache.commons.lang3.ObjectUtils;
 import org.ccjmne.orca.api.modules.Restrictions;
 import org.ccjmne.orca.api.rest.fetch.ResourcesEndpoint;
 import org.ccjmne.orca.api.utils.Constants;
@@ -34,7 +26,6 @@ import org.ccjmne.orca.api.utils.SafeDateFormat;
 import org.ccjmne.orca.api.utils.Transactions;
 import org.ccjmne.orca.jooq.classes.Sequences;
 import org.ccjmne.orca.jooq.classes.tables.records.SitesEmployeesRecord;
-import org.ccjmne.orca.jooq.classes.tables.records.SitesTagsRecord;
 import org.jooq.DSLContext;
 import org.jooq.InsertValuesStep3;
 import org.jooq.TableField;
@@ -55,58 +46,6 @@ public class UpdateEndpoint {
 		}
 
 		this.ctx = ctx;
-	}
-
-	@PUT
-	@Path("sites/{site_pk}")
-	@Consumes(MediaType.APPLICATION_JSON)
-	@SuppressWarnings("unchecked")
-	public Boolean insertSite(@PathParam("site_pk") final Integer site_pk, final Map<String, Object> siteDefinition) {
-		return Transactions.with(this.ctx, transactionCtx -> {
-			final Map<String, Object> tags = ObjectUtils.defaultIfNull((Map<String, Object>) siteDefinition.remove("tags"), Collections.emptyMap());
-			final boolean exists = transactionCtx.fetchExists(SITES, SITES.SITE_PK.eq(site_pk));
-			if (exists) {
-				transactionCtx.update(SITES).set(siteDefinition).where(SITES.SITE_PK.eq(site_pk)).execute();
-			} else {
-				transactionCtx.insertInto(SITES)
-						.set(siteDefinition)
-						.set(SITES.SITE_PK, site_pk)
-						.execute();
-			}
-
-			transactionCtx.deleteFrom(SITES_TAGS).where(SITES_TAGS.SITA_SITE_FK.eq(site_pk)).execute();
-			transactionCtx.batchInsert(tags.entrySet().stream()
-					.map(tag -> {
-						if (Constants.TAGS_VALUE_NONE.equals(tag.getValue()) || Constants.TAGS_VALUE_UNIVERSAL.equals(tag.getValue())) {
-							throw new IllegalArgumentException(String.format("Invalid tag value: '%s'", tag.getValue()));
-						}
-
-						return new SitesTagsRecord(site_pk, Integer.valueOf(tag.getKey()), String.valueOf(tag.getValue()));
-					}).collect(Collectors.toList())).execute();
-			return Boolean.valueOf(!exists);
-		});
-	}
-
-	/**
-	 * @return <code>true</code> iff that site was actually existing
-	 */
-	@DELETE
-	@Path("sites/{site_pk}")
-	public Boolean deleteSite(@PathParam("site_pk") final Integer site_pk) {
-		// Database CASCADEs the deletion of linked users, if any
-		// Database CASCADEs the deletion of its tags
-		return Transactions.with(this.ctx, transactionCtx -> {
-			// Set deleted entity employees' site to UNASSIGNED
-			final boolean exists = transactionCtx.selectFrom(SITES).where(SITES.SITE_PK.equal(site_pk)).fetch().isNotEmpty();
-			if (exists) {
-				transactionCtx.update(SITES_EMPLOYEES)
-						.set(SITES_EMPLOYEES.SIEM_SITE_FK, Constants.DECOMMISSIONED_SITE)
-						.where(SITES_EMPLOYEES.SIEM_SITE_FK.eq(site_pk)).execute();
-				transactionCtx.delete(SITES).where(SITES.SITE_PK.eq(site_pk)).execute();
-			}
-
-			return Boolean.valueOf(exists);
-		});
 	}
 
 	@POST
