+++ src/main/webapp/WEB-INF/web.xml
@@ -153,11 +153,12 @@
 			<url-pattern>/certificates/*</url-pattern>
 			<url-pattern>/client/*</url-pattern>
 			<url-pattern>/employees-notes/*</url-pattern>
+			<url-pattern>/sites/*</url-pattern>
 			<url-pattern>/tags/*</url-pattern>
 			<url-pattern>/update/*</url-pattern>
 			<url-pattern>/users-admin/*</url-pattern>
 
-			<http-method>GET</http-method>
+			<http-method>GET</http-method><!-- specific to users-admin -->
 			<http-method>DELETE</http-method>
 			<http-method>POST</http-method>
 			<http-method>PUT</http-method>
